requirements.txt
amqp==2.4.2
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
Django==1.11.18
django-celery==3.2.2
django-model-utils==3.1.2
djangorestframework==3.9.2
kombu==4.5.0
psycopg2==2.7.7
pytz==2019.1
redis==3.2.1
vine==1.3.0
